requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
sqlparse==0.4.4
tzdata==2024.1
